Define a Pipeline behaviour that our individual pipelines implement
Rename the root function to init.

Rename the existing module to be PipelineHelpers.
