+++ test/concentrate/pipeline/stop_time_updates_pipeline_test.exs
@@ -2,7 +2,7 @@ defmodule Concentrate.Pipeline.StopTimeUpdatesPipelineTest do
   use ExUnit.Case, async: true
   import Test.Support.Helpers
 
-  describe "pipeline/1" do
+  describe "init/1" do
     setup do
       reassign_env(:realtime, :trip_fn, fn _trip_id -> nil end)
       reassign_env(:realtime, :block_fn, fn _block_id, _service_id -> nil end)
@@ -13,7 +13,7 @@ defmodule Concentrate.Pipeline.StopTimeUpdatesPipelineTest do
         trip_updates_url: "http://example.com/TripUpdates_enhanced.json"
       ]
 
-      pipeline_elements = Concentrate.Pipeline.StopTimeUpdatesPipeline.pipeline(opts)
+      pipeline_elements = Concentrate.Pipeline.StopTimeUpdatesPipeline.init(opts)
 
       assert length(pipeline_elements) == 2
     end
