+++ test/concentrate/pipeline/vehicle_positions_pipeline_test.exs
@@ -2,7 +2,7 @@ defmodule Concentrate.Pipeline.VehiclePositionsPipelineTest do
   use ExUnit.Case, async: true
   import Test.Support.Helpers
 
-  describe "pipeline/1" do
+  describe "init/1" do
     setup do
       reassign_env(:realtime, :trip_fn, fn _trip_id -> nil end)
       reassign_env(:realtime, :block_fn, fn _block_id, _service_id -> nil end)
@@ -15,7 +15,7 @@ defmodule Concentrate.Pipeline.VehiclePositionsPipelineTest do
         swiftly_realtime_vehicles_url: "http://example.com/swiftly_realtime_vehicles.json"
       ]
 
-      pipeline_elements = Concentrate.Pipeline.VehiclePositionsPipeline.pipeline(opts)
+      pipeline_elements = Concentrate.Pipeline.VehiclePositionsPipeline.init(opts)
 
       assert length(pipeline_elements) == 4
     end
